fixed some issues with unsigned to signed conversion

+++ code/connect.h
@@ -24,7 +24,7 @@ const unsigned int PROMPT_LIFEFORCE         = (1<<17);
 const unsigned int PROMPT_TIME              = (1<<18);
 // Add new prompt options here.
 const unsigned int PROMPT_CLASSIC_ANSIBAR   = (1<<30);
-const unsigned int PROMPT_CLIENT_PROMPT     = (1<<31);
+const unsigned int PROMPT_CLIENT_PROMPT     = (unsigned)(1<<31);
 
 const int MAX_TRAITS=16;
